+++ out/extension.js
@@ -49,11 +49,19 @@ const session_end_1 = __importDefault(require("./email/session_end"));
 // This method is called when extension is activated
 // extension is activated the very first time the command is executed
 function activate(context) {
+    //  Updating the settings
+    if (vscode.workspace.getConfiguration("work-progress").get("apiKey") !== context.globalState.get("apiKey")) {
+        context.globalState.update("apiKey", vscode.workspace.getConfiguration("work-progress").get("apiKey"));
+    }
+    // Checking if the user is logged in
+    if (vscode.workspace.getConfiguration("work-progress").get("apiKey") === undefined || vscode.workspace.getConfiguration("work-progress").get("apiKey") === "" || context.globalState.get("apiKey") === undefined || context.globalState.get("apiKey") === "") {
+        vscode.window.showInformationMessage("Please log in to Work Progress to use email notifications. You can do it on https://vswork-progress.vercel.app or https://andrinoff.github.io/workprogress/");
+    }
     // This log helps with cathing errors at the startup
     console.log('Congratulations, your extension "work-progress" is now active!');
-    // Check how long the person has been working
+    // Send the email with the time worked
     (0, check_screen_time_1.default)(context);
-    if (context.globalState.get("time_worked") !== undefined && context.globalState.get("time_worked") !== 0 && vscode.workspace.getConfiguration("work-progress").get("session", false)) {
+    if (context.globalState.get("time_worked") !== undefined && context.globalState.get("time_worked") !== "0" && vscode.workspace.getConfiguration("work-progress").get("session", false)) {
         console.log("sending email for time worked: " + context.globalState.get("time_worked"));
         (0, session_end_1.default)(context, parseInt(context.globalState.get("time_worked") || "0") / 60);
         context.globalState.update("time_worked", "0");