+++ CHANGELOG.md
@@ -14,3 +14,14 @@ All notable changes to the "work-progress" extension will be documented in this 
 - Emails sent after each session (have to enable, sorry)
 - Time measured without having to enter commands
 - API is now manageble from settings
+
+## 1.1.0 
+
+- Checks idle time
+- Sends email if user not focused for 1 hour and stops the timer
+- Updates user settings on activation
+
+## 1.1.1 (released)
+
+- Emails turned off (temporarly)
+- Not fixed idle and focused time